requirements.txt
alabaster==0.7.12
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
Brian2==2.3.0.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.21
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
docopt==0.6.2
docutils==0.16
entrypoints==0.3
experimentator==0.3.1
flake8==3.8.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
h5py==2.10.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.7.0
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lazyarray==0.3.3
-e git+https://github.com/chanokin/L2L@6149fbd752572808bf9727176ece78ba6f18270f#egg=Learning_to_Learn
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mpi4py==3.0.3
mpmath==1.1.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
neo==0.6.1
networkx==2.4
notebook==6.0.3
numpy==1.19.0
opencv-python==4.4.0.40
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py-cpuinfo==7.0.0
pycairo==1.19.1
pycodestyle==2.6.0
pyflakes==2.2.0
-e git+https://github.com/genn-team/genn@20406d9e892634f186048473f2f7f59dc5f96515#egg=pygenn
Pygments==2.6.1
PyGObject==3.36.1
pylint==2.5.3
PyNN==0.9.2
-e git+https://github.com/genn-team/pynn_genn@6f7baa92cf345f2231f3076ba8da8f6339aaec6a#egg=pynn_genn
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqtgraph==0.11.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
quantities==0.12.4
requests==2.24.0
sacred==0.8.1
schema==0.7.2
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
Send2Trash==1.5.0
sentinel==0.1.2
six==1.15.0
sklearn==0.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxjp.themes.basicstrap==0.5.0
sympy==1.6.2
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
